+++ voxygen/src/audio/sfx/event_mapper/movement/mod.rs
@@ -101,12 +101,15 @@ impl EventMapper for MovementEventMapper {
                 };
 
                 // Check for SFX config entry for this movement
-                if Self::should_emit(internal_state, triggers.get_key_value(&mapped_event)) {
+                if Self::should_emit(internal_state, triggers.get_key_value(&mapped_event))
+                    && let Some(player_pos) = state.read_component_copied::<Pos>(player_entity)
+                {
                     let sfx_trigger_item = triggers.get_key_value(&mapped_event);
                     audio.emit_sfx(
                         sfx_trigger_item,
                         pos.0,
                         Some(Self::get_volume_for_body_type(body)),
+                        player_pos.0,
                     );
                     internal_state.time = Instant::now();
                     internal_state.steps_taken = 0.0;